+++ apps/meteor/client/views/marketplace/helpers/filterAppsByCategories.spec.ts
@@ -1,8 +1,5 @@
-/* eslint-env mocha */
-import { expect } from 'chai';
-
-import { filterAppsByCategories } from '../../../../../../../client/views/marketplace/helpers/filterAppsByCategories';
-import type { App } from '../../../../../../../client/views/marketplace/types';
+import type { App } from '../types';
+import { filterAppsByCategories } from './filterAppsByCategories';
 
 describe('filterAppsByCategories', () => {
 	it('should return true if the app is in the categories', () => {
@@ -11,14 +8,15 @@ describe('filterAppsByCategories', () => {
 		};
 		const categories = ['category1'];
 		const result = filterAppsByCategories(app as App, categories);
-		expect(result).to.be.true;
+		expect(result).toBe(true);
 	});
+
 	it('should return false if the app is not in the categories', () => {
 		const app = {
 			categories: ['category1', 'category2'],
 		};
 		const categories = ['category3'];
 		const result = filterAppsByCategories(app as App, categories);
-		expect(result).to.be.false;
+		expect(result).toBe(false);
 	});
 });
