+++ apps/meteor/client/views/marketplace/helpers/filterAppsByFree.spec.ts
@@ -1,8 +1,6 @@
-/* eslint-env mocha */
 import type { PurchaseType } from '@rocket.chat/core-typings';
-import { expect } from 'chai';
 
-import { filterAppsByFree } from '../../../../../../../client/views/marketplace/helpers/filterAppsByFree';
+import { filterAppsByFree } from './filterAppsByFree';
 
 describe('filterAppsByFree', () => {
 	it('should return true if app purchase type is buy and price does not exist or is 0', () => {
@@ -12,8 +10,9 @@ describe('filterAppsByFree', () => {
 			price: 0,
 		};
 		const result = filterAppsByFree(app);
-		expect(result).to.be.true;
+		expect(result).toBe(true);
 	});
+
 	it('should return false if app purchase type is not buy', () => {
 		const purchaseType: PurchaseType = 'subscription';
 		const app = {
@@ -21,8 +20,9 @@ describe('filterAppsByFree', () => {
 			price: 0,
 		};
 		const result = filterAppsByFree(app);
-		expect(result).to.be.false;
+		expect(result).toBe(false);
 	});
+
 	it('should return false if app price exists and is different than 0', () => {
 		const purchaseType: PurchaseType = 'buy';
 		const app = {
@@ -30,8 +30,9 @@ describe('filterAppsByFree', () => {
 			price: 5,
 		};
 		const result = filterAppsByFree(app);
-		expect(result).to.be.false;
+		expect(result).toBe(false);
 	});
+
 	it('should return false if both app purchase type is different than buy and price exists and is different than 0', () => {
 		const purchaseType: PurchaseType = 'subscription';
 		const app = {
@@ -39,6 +40,6 @@ describe('filterAppsByFree', () => {
 			price: 5,
 		};
 		const result = filterAppsByFree(app);
-		expect(result).to.be.false;
+		expect(result).toBe(false);
 	});
 });
