+++ apps/meteor/client/views/marketplace/helpers/filterAppsByPaid.spec.ts
@@ -1,8 +1,6 @@
-/* eslint-env mocha */
 import type { PurchaseType } from '@rocket.chat/core-typings';
-import { expect } from 'chai';
 
-import { filterAppsByPaid } from '../../../../../../../client/views/marketplace/helpers/filterAppsByPaid';
+import { filterAppsByPaid } from './filterAppsByPaid';
 
 describe('filterAppsByPaid', () => {
 	it('should return true if both app purchase type is subscription and app price exists and is not 0', () => {
@@ -12,8 +10,9 @@ describe('filterAppsByPaid', () => {
 			price: 5,
 		};
 		const result = filterAppsByPaid(app);
-		expect(result).to.be.true;
+		expect(result).toBe(true);
 	});
+
 	it('should return true if app purchase type is subscription', () => {
 		const purchaseType: PurchaseType = 'subscription';
 		const app = {
@@ -21,8 +20,9 @@ describe('filterAppsByPaid', () => {
 			price: 0,
 		};
 		const result = filterAppsByPaid(app);
-		expect(result).to.be.true;
+		expect(result).toBe(true);
 	});
+
 	it('should return true if app price exists and is not 0', () => {
 		const purchaseType: PurchaseType = 'buy';
 		const app = {
@@ -30,8 +30,9 @@ describe('filterAppsByPaid', () => {
 			price: 5,
 		};
 		const result = filterAppsByPaid(app);
-		expect(result).to.be.true;
+		expect(result).toBe(true);
 	});
+
 	it('should return false if both app price does not exist or is 0 and app purchase type is not subscription', () => {
 		const purchaseType: PurchaseType = 'buy';
 		const app = {
@@ -39,6 +40,6 @@ describe('filterAppsByPaid', () => {
 			price: 0,
 		};
 		const result = filterAppsByPaid(app);
-		expect(result).to.be.false;
+		expect(result).toBe(false);
 	});
 });
