+++ apps/meteor/client/views/marketplace/helpers/filterAppsByText.spec.ts
@@ -1,7 +1,4 @@
-/* eslint-env mocha */
-import { expect } from 'chai';
-
-import { filterAppsByText } from '../../../../../../../client/views/marketplace/helpers/filterAppsByText';
+import { filterAppsByText } from './filterAppsByText';
 
 describe('filterAppsByText', () => {
 	it('should return true if the text is the name of an app', () => {
@@ -10,14 +7,15 @@ describe('filterAppsByText', () => {
 		};
 		const text = 'name1';
 		const result = filterAppsByText(app.name, text);
-		expect(result).to.be.true;
+		expect(result).toBe(true);
 	});
+
 	it('should return false if the text is not the name of an app', () => {
 		const app = {
 			name: 'name1',
 		};
 		const text = 'name2';
 		const result = filterAppsByText(app.name, text);
-		expect(result).to.be.false;
+		expect(result).toBe(false);
 	});
 });
