+++ apps/meteor/client/views/marketplace/helpers/sortAppsByAlphabeticalOrInverseOrder.spec.ts
@@ -1,7 +1,4 @@
-/* eslint-env mocha */
-import { expect } from 'chai';
-
-import { sortAppsByAlphabeticalOrInverseOrder } from '../../../../../../../client/views/marketplace/helpers/sortAppsByAlphabeticalOrInverseOrder';
+import { sortAppsByAlphabeticalOrInverseOrder } from './sortAppsByAlphabeticalOrInverseOrder';
 
 describe('sortAppsByAlphabeticalOrder', () => {
 	it.skip('should return a positive number if first word is, alphabetically, after second word', () => {
@@ -10,22 +7,24 @@ describe('sortAppsByAlphabeticalOrder', () => {
 
 		const result = sortAppsByAlphabeticalOrInverseOrder(firstWord, secondWord);
 
-		expect(result).to.be.above(0);
+		expect(result).toBeGreaterThan(0);
 	});
+
 	it.skip('should return a negative number if first word is, alphabetically, before second word', () => {
 		const firstWord = 'Bravo';
 		const secondWord = 'Alfa';
 
 		const result = sortAppsByAlphabeticalOrInverseOrder(firstWord, secondWord);
 
-		expect(result).to.be.below(0);
+		expect(result).toBeLessThan(0);
 	});
+
 	it('should return 0 if the words are the equivalent', () => {
 		const firstWord = 'Alfa';
 		const secondWord = 'Alfa';
 
 		const result = sortAppsByAlphabeticalOrInverseOrder(firstWord, secondWord);
 
-		expect(result).to.be.equal(0);
+		expect(result).toBe(0);
 	});
 });
